feat: improve extensibility of categories page
